Update Example to RN 0.34 and Xcode 8
