+++ Example/android/settings.gradle
@@ -2,4 +2,4 @@ rootProject.name = 'Example'
 
 include ':app'
 include ':react-native-camera'
-project(':react-native-camera').projectDir = new File(rootProject.projectDir, '../../android')
+project(':react-native-camera').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-camera/android')
