+++ Example/package.json
@@ -6,8 +6,8 @@
     "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
-    "react": "15.0.2",
-    "react-native": "^0.26.2",
+    "react": "~15.3.0",
+    "react-native": "^0.34.0",
     "react-native-camera": "file:../"
   }
 }
